feat(recorder): change styling of feed/add-page
Updated the recorder button to be larger (24x24) and added dynamic text to display 'Stop' when recor
ding and 'Rec' otherwise. Also added a spin animation when recording.

Increased the minimum height of the button and updated text styling to be larger and bold. The Voice
Memo component is now conditionally rendered based on the visibility of the textarea.