feat(pgflow-worker): utilize queryObject for queue creation
Updated the pgflow-worker to use the new `queryObject` function for creating a queue with `pgmq.crea
te('yolo')`. This change simplifies the code and leverages the new functionality provided by the use
ConnectionPool module.

also add queryObject and queryArray functions to useConnectionPool

Added `queryObject` and `queryArray` functions to the `useConnectionPool` module to facilitate query
ing the database with object and array results. This enhances the flexibility of database interactio
ns within the pgflow context.